Auth / User - Héros (unique) / User - Site Héros V1.3.0
Nom du Site : Le Site du Héros
Version : 1.3.0

Ajout système Authentification avec compte User ( Login / Register / Logout )
Possibilité de créer / modifier ou supprimer que l'héros (unique) par rapport au compte User
- migrations
- models / request / controllers / views
- route "web.php" + middleware

+++ app/Http/Controllers/HerosController.php
@@ -6,6 +6,7 @@ use App\Http\Requests\HerosRequest;
 use App\Models\Heros;
 use App\Models\Incidents;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class HerosController extends Controller
 {
@@ -36,6 +37,13 @@ class HerosController extends Controller
     {
         $incidents = Incidents::all();
 
+        $user = Auth::user();
+        $hasHero = $user->heros()->exists();
+
+        if ($hasHero) {
+            return redirect()->route('heros.index')->with('status-danger', 'Vous ne pouvez créer qu\'un seul héros.');
+        }
+
         return view('heros.form', compact('incidents'));
     }
 
@@ -48,7 +56,7 @@ class HerosController extends Controller
         $selectedIncidents = $request->input('incidents', []);
         $heros->incidents()->attach($selectedIncidents);
 
-        return redirect()->route('heros.index')->with('success', 'Héros créé avec succès');
+        return redirect()->route('heros.show', $heros)->with('status-success', 'Héros créé avec succès');
     }
 
     public function show(Heros $hero)
@@ -58,6 +66,10 @@ class HerosController extends Controller
 
     public function edit(Heros $hero)
     {
+        if ($hero->user_id !== Auth::id()) {
+            return redirect()->route('heros.index')->with('status-danger', 'Vous n\'êtes pas autorisé à modifier ce héros.');
+        }
+
         $incidents = Incidents::all();
 
         return view('heros.form', compact('hero', 'incidents'));
@@ -65,6 +77,11 @@ class HerosController extends Controller
 
     public function update(HerosRequest $request, Heros $hero)
     {
+
+        if ($hero->user_id !== Auth::id()) {
+            return redirect()->route('heros.index')->with('status-danger', 'Vous n\'êtes pas autorisé à mettre à jour ce héros.');
+        }
+
         $herosData = $request->validated();
 
         $hero->update($herosData);
@@ -72,14 +89,20 @@ class HerosController extends Controller
         $selectedIncidents = $request->input('incidents', []);
         $hero->incidents()->sync($selectedIncidents);
 
-        return redirect()->route('heros.index')->with('success', 'Héros mis à jour avec succès');
+        return redirect()->route('heros.show', $hero)->with('status-success', 'Héros mis à jour avec succès');
     }
 
     public function destroy(Heros $hero)
     {
+
+        if ($hero->user_id !== Auth::id()) {
+            return redirect()->route('heros.index')->with('status-danger', 'Vous n\'êtes pas autorisé à supprimer ce héros.');
+        }
+
         $hero->incidents()->detach();
         $hero->delete();
 
-        return redirect()->route('heros.index')->with('success', 'Héros supprimé avec succès');
+        return redirect()->route('heros.index')->with('status-success', 'Héros supprimé avec succès');
     }
+
 }
